moved workflow to its own module
