+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_v2/__init__.py
@@ -1,7 +1,8 @@
-from ._action import Action, Workflow
+from ._action import Action
 from ._marker import mark_step
 from ._play_context import PlayContext
 from ._player import PlayerManager
+from ._workflow import Workflow
 
 __all__: tuple[str, ...] = (
     "mark_step",
