+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_v2/_action.py
@@ -2,10 +2,6 @@ from typing import Any, Callable, Optional
 
 from pydantic import BaseModel, Field, validator
 
-from ._errors import (
-    NextActionNotInPlayCatalogException,
-    OnErrorActionNotInPlayCatalogException,
-)
 from ._models import ActionName
 
 
@@ -55,35 +51,3 @@ class Action(BaseModel):
                     f"`{step.__name__}.return_type` should be of type {dict[str, Any]}"
                 )
         return steps
-
-
-class Workflow:
-    """contains Action entries which define links to `next_action` and `on_error_action`"""
-
-    def __init__(self, *actions: Action) -> None:
-        self._registry: dict[ActionName, Action] = {s.name: s for s in actions}
-        for action in actions:
-            if (
-                action.on_error_action is not None
-                and action.on_error_action not in self._registry
-            ):
-                raise OnErrorActionNotInPlayCatalogException(
-                    action_name=action.name,
-                    on_error_action=action.on_error_action,
-                    workflow=self._registry,
-                )
-            if (
-                action.next_action is not None
-                and action.next_action not in self._registry
-            ):
-                raise NextActionNotInPlayCatalogException(
-                    action_name=action.name,
-                    next_action=action.next_action,
-                    workflow=self._registry,
-                )
-
-    def __contains__(self, item: ActionName) -> bool:
-        return item in self._registry
-
-    def __getitem__(self, key: ActionName) -> Action:
-        return self._registry[key]
